lab-3-vecSearch/requirements.txt
numpy>=1.24
faiss-cpu>=1.7.4
tqdm>=4.66
requests>=2.31
matplotlib>=3.7
